+++ core/src/main/java/com/vanpra/composematerialdialogs/MaterialDialogButtons.kt
@@ -20,6 +20,7 @@ import androidx.compose.ui.layout.layoutId
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import java.util.Locale
 
@@ -112,6 +113,7 @@ class MaterialDialogButtons(private val dialog: MaterialDialog) {
     @Composable
     fun button(
         text: String? = null,
+        textStyle: TextStyle = MaterialTheme.typography.button,
         @StringRes res: Int? = null,
         onClick: () -> Unit = {}
     ) {
@@ -124,7 +126,7 @@ class MaterialDialogButtons(private val dialog: MaterialDialog) {
                 .layoutId(MaterialDialogButtonTypes.Text)
                 .testTag(MaterialDialogButtonTypes.Text.testTag),
         ) {
-            Text(text = buttonText, style = MaterialTheme.typography.button)
+            Text(text = buttonText, style = textStyle)
         }
     }
 
@@ -140,6 +142,7 @@ class MaterialDialogButtons(private val dialog: MaterialDialog) {
     @Composable
     fun positiveButton(
         text: String? = null,
+        textStyle: TextStyle = MaterialTheme.typography.button,
         @StringRes res: Int? = null,
         disableDismiss: Boolean = false,
         onClick: () -> Unit = {}
@@ -164,7 +167,7 @@ class MaterialDialogButtons(private val dialog: MaterialDialog) {
                 .testTag(MaterialDialogButtonTypes.Positive.testTag),
             enabled = buttonEnabled && dialog.positiveButtonEnabledOverride
         ) {
-            Text(text = buttonText, style = MaterialTheme.typography.button)
+            Text(text = buttonText, style = textStyle)
         }
     }
 
@@ -179,6 +182,7 @@ class MaterialDialogButtons(private val dialog: MaterialDialog) {
     @Composable
     fun negativeButton(
         text: String? = null,
+        textStyle: TextStyle = MaterialTheme.typography.button,
         @StringRes res: Int? = null,
         onClick: () -> Unit = {}
     ) {
@@ -195,7 +199,7 @@ class MaterialDialogButtons(private val dialog: MaterialDialog) {
             modifier = Modifier.layoutId(MaterialDialogButtonTypes.Negative)
                 .testTag(MaterialDialogButtonTypes.Negative.testTag),
         ) {
-            Text(text = buttonText, style = MaterialTheme.typography.button)
+            Text(text = buttonText, style = textStyle)
         }
     }
 
@@ -206,7 +210,11 @@ class MaterialDialogButtons(private val dialog: MaterialDialog) {
      * @param onClick a callback which is called when the button is pressed
      */
     @Composable
-    fun accessibilityButton(icon: ImageVector, onClick: () -> Unit) {
+    fun accessibilityButton(
+        icon: ImageVector,
+        colorFilter: ColorFilter = ColorFilter.tint(MaterialTheme.colors.onBackground),
+        onClick: () -> Unit
+    ) {
         Box(
             Modifier
                 .size(48.dp)
@@ -219,7 +227,7 @@ class MaterialDialogButtons(private val dialog: MaterialDialog) {
                 icon,
                 contentDescription = null,
                 modifier = Modifier.size(24.dp),
-                colorFilter = ColorFilter.tint(MaterialTheme.colors.onBackground)
+                colorFilter = colorFilter
             )
         }
     }