+++ datetime/src/androidTest/java/com/vanpra/composematerialdialogs/datetime/test/screenshot/DateTimePickerTest.kt
@@ -3,12 +3,13 @@ package com.vanpra.composematerialdialogs.datetime.test.screenshot
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.performClick
 import com.karumi.shot.ScreenshotTest
-import com.vanpra.composematerialdialogs.datetime.datepicker
+import com.vanpra.composematerialdialogs.datetime.date.datepicker
 import com.vanpra.composematerialdialogs.datetime.datetimepicker
-import com.vanpra.composematerialdialogs.datetime.timepicker.timepicker
+import com.vanpra.composematerialdialogs.datetime.time.timepicker
 import com.vanpra.composematerialdialogs.test.utils.DialogWithContent
 import com.vanpra.composematerialdialogs.test.utils.extensions.onDialog
-import com.vanpra.composematerialdialogs.test.utils.extensions.onPositiveButton
+import com.vanpra.composematerialdialogs.test.utils.extensions.onDialogDateNextMonth
+import com.vanpra.composematerialdialogs.test.utils.extensions.onDialogDateSelector
 import com.vanpra.composematerialdialogs.test.utils.extensions.setContentAndWaitForIdle
 import org.junit.Rule
 import org.junit.Test
@@ -63,7 +64,20 @@ class DateTimePickerTest : ScreenshotTest {
             }
         }
         compareScreenshot(composeTestRule.onDialog())
-        composeTestRule.onPositiveButton().performClick()
+    }
+
+    @Test
+    fun datetimePickerSelectionOnNextMonth() {
+        composeTestRule.setContentAndWaitForIdle {
+            DialogWithContent {
+                datetimepicker(
+                    initialDateTime = LocalDateTime.of(2021, 1, 1, 12, 0)
+                )
+            }
+        }
+
+        composeTestRule.onDialogDateNextMonth().performClick()
+        composeTestRule.onDialogDateSelector(3).performClick()
         compareScreenshot(composeTestRule.onDialog())
     }
 }